Extract RowGroupOrderOptions to a light header
table_function.hpp (included by ~1080 TUs) pulled the heavy
storage/table/row_group_reorderer.hpp solely to name RowGroupOrderOptions in one
typedef (unique_ptr<RowGroupOrderOptions> by pointer). row_group_reorderer.hpp
drags row_group.hpp + row_group_segment_tree.hpp + segment_tree.hpp and their
closures onto every includer of table_function.hpp.

Move RowGroupOrderOptions (and the OrderByStatistics/OrderByColumnType enums it
uses) into a new light header storage/table/row_group_order_options.hpp that
pulls only storage_index + order_type + optional_idx. table_function.hpp now
forward-declares RowGroupOrderOptions and drops the heavy include; the genuine
consumers (logical_get, table_scan, row_group_pruner) include the light header;
row_group_reorderer.hpp re-includes it so RowGroupReorderer is unchanged.

Removing row_group_reorderer.hpp from table_function's closure stopped it from
transitively forward-declaring a set of storage/checkpoint types for headers
that had relied on that path; those now declare what they use directly:
- storage/table/column_data.hpp: fwd-declare PartialBlockManager, DuckTableEntry,
  ColumnCheckpointState, ColumnSegmentInfo, ColumnSegmentInfoScanOptions;
  include enums/scan_options.hpp (UpdateScanType used by value)
- storage/table/data_table_info.hpp: fwd-declare CheckpointOptions
- storage/checkpoint/table_data_writer.hpp: fwd-declare RowGroupCollection,
  RowGroupPointer

total_tu_include_cost
before: 426,102
after:  411,842
